Process Activity Window Support in Mobile Mode Migration
